+++ tkcalendar/calendar_.py
@@ -68,8 +68,11 @@ class Calendar(ttk.Frame):
         firstweekday : "monday" or "sunday"
             first day of the week
 
-        showweeknumbers : boolean (default is True)
+        showweeknumbers : bool (default is True)
             whether to display week numbers.
+
+        showothermonthdays : bool (default is True)
+            whether to display the last days of the previous month and the first of the next month.
 
         locale : str
             locale to use, e.g. 'en_US'
@@ -261,6 +264,7 @@ class Calendar(ttk.Frame):
                    'textvariable',
                    'locale',
                    'showweeknumbers',
+                   'showothermonthdays',
                    'firstweekday',
                    'selectbackground',
                    'selectforeground',
@@ -300,6 +304,7 @@ class Calendar(ttk.Frame):
                             'textvariable': self._textvariable,
                             'firstweekday': firstweekday,
                             'showweeknumbers': showweeknumbers,
+                            'showothermonthdays': kw.pop('showothermonthdays', True),
                             'selectbackground': active_bg,
                             'selectforeground': 'white',
                             'disabledselectbackground': dis_active_bg,
@@ -563,6 +568,8 @@ class Calendar(ttk.Frame):
             elif key is "tooltipdelay":
                 self.tooltip_wrapper.configure(delay=value)
             self._properties[key] = value
+            if key is 'showothermonthdays':
+                self._display_calendar()
 
     def _textvariable_trace(self, *args):
         """Connect StringVar value with selected date."""
@@ -668,11 +675,62 @@ class Calendar(ttk.Frame):
         self._header_month.configure(text=header.title())
         self._header_year.configure(text=str(year))
 
-        # update calendar shown dates
-        cal = self._cal.monthdatescalendar(year, month)
-
         # remove previous tooltips
         self.tooltip_wrapper.remove_all()
+
+        # update calendar shown dates
+        if self['showothermonthdays']:
+            self._display_days_with_othermonthdays()
+        else:
+            self._display_days_without_othermonthdays()
+
+        self._display_selection()
+
+    def _display_days_without_othermonthdays(self):
+        year, month = self._date.year, self._date.month
+
+        cal = self._cal.monthdays2calendar(year, month)
+        while len(cal) < 6:
+            cal.append([(0, i) for i in range(7)])
+
+        week_days = {i: 'normal.%s.TLabel' % self._style_prefixe for i in range(7)}  # style names depending on the type of day
+        offset = (self['firstweekday'] == 'sunday') * 6
+        week_days[(5 - offset) % 7] = 'we.%s.TLabel' % self._style_prefixe
+        week_days[(6 - offset) % 7] = 'we.%s.TLabel' % self._style_prefixe
+        _, week_nb, d = self._date.isocalendar()
+        if d == 7 and self['firstweekday'] == 'sunday':
+            week_nb += 1
+        modulo = max(week_nb, 52)
+        for i_week in range(6):
+            if i_week == 0 or cal[i_week][0][0]:
+                self._week_nbs[i_week].configure(text=str((week_nb + i_week - 1) % modulo + 1))
+            else:
+                self._week_nbs[i_week].configure(text='')
+            for i_day in range(7):
+                day_number, week_day = cal[i_week][i_day]
+                style = week_days[i_day]
+                label = self._calendar[i_week][i_day]
+                if day_number:
+                    txt = str(day_number)
+                    label.configure(text=txt, style=style)
+                    date = self.date(year, month, day_number)
+                    if date in self._calevent_dates:
+                        ev_ids = self._calevent_dates[date]
+                        i = len(ev_ids) - 1
+                        while i >= 0 and not self.calevents[ev_ids[i]]['tags']:
+                            i -= 1
+                        if i >= 0:
+                            tag = self.calevents[ev_ids[i]]['tags'][-1]
+                            label.configure(style='tag_%s.%s.TLabel' % (tag, self._style_prefixe))
+                        text = '\n'.join(['➢ {}'.format(self.calevents[ev]['text']) for ev in ev_ids])
+                        self.tooltip_wrapper.add_tooltip(label, text)
+                else:
+                    label.configure(text='', style=style)
+
+    def _display_days_with_othermonthdays(self):
+        year, month = self._date.year, self._date.month
+
+        cal = self._cal.monthdatescalendar(year, month)
 
         next_m = month + 1
         y = year
@@ -688,8 +746,7 @@ class Calendar(ttk.Frame):
             if len(cal) < 6:
                 cal.append(self._cal.monthdatescalendar(y, next_m)[i + 1])
 
-        # style names depending on the type of day
-        week_days = {i: 'normal' for i in range(7)}
+        week_days = {i: 'normal' for i in range(7)}  # style names depending on the type of day
         offset = (self['firstweekday'] == 'sunday') * 6
         week_days[(5 - offset) % 7] = 'we'
         week_days[(6 - offset) % 7] = 'we'
@@ -718,8 +775,6 @@ class Calendar(ttk.Frame):
                         label.configure(style='tag_%s.%s.TLabel' % (tag, self._style_prefixe))
                     text = '\n'.join(['➢ {}'.format(self.calevents[ev]['text']) for ev in ev_ids])
                     self.tooltip_wrapper.add_tooltip(label, text)
-
-        self._display_selection()
 
     def _get_day_coords(self, date):
         year = date.year